refactor: transfer transaction and airdrop transaction
