+++ src/account/TransferTransaction.js
@@ -24,16 +24,12 @@ import AccountId from "./AccountId.js";
 import Transaction, {
     TRANSACTION_REGISTRY,
 } from "../transaction/Transaction.js";
-import Long from "long";
-import NullableTokenDecimalMap from "./NullableTokenDecimalMap.js";
 import Transfer from "../Transfer.js";
 import TokenTransfer from "../token/TokenTransfer.js";
-import TokenTransferMap from "./TokenTransferMap.js";
 import HbarTransferMap from "./HbarTransferMap.js";
-import TokenNftTransferMap from "./TokenNftTransferMap.js";
-import TokenTransferAccountMap from "./TokenTransferAccountMap.js";
 import TokenNftTransfer from "../token/TokenNftTransfer.js";
 import NftId from "../token/NftId.js";
+import AbstractTokenTransfer from "../token/AbstractTokenTransfer.js";
 
 /**
  * @typedef {import("../long.js").LongObject} LongObject
@@ -48,10 +44,6 @@ import NftId from "../token/NftId.js";
  * @typedef {import("@hashgraph/proto").proto.ITransactionBody} HashgraphProto.proto.ITransactionBody
  * @typedef {import("@hashgraph/proto").proto.ITransactionResponse} HashgraphProto.proto.ITransactionResponse
  * @typedef {import("@hashgraph/proto").proto.ICryptoTransferTransactionBody} HashgraphProto.proto.ICryptoTransferTransactionBody
- * @typedef {import("@hashgraph/proto").proto.ITokenID} HashgraphProto.proto.ITokenID
- * @typedef {import("@hashgraph/proto").proto.IAccountID} HashgraphProto.proto.IAccountID
- * @typedef {import("@hashgraph/proto").proto.IAccountAmount} HashgraphProto.proto.IAccountAmount
- * @typedef {import("@hashgraph/proto").proto.ITokenTransferList} HashgraphProto.proto.ITokenTransferList
  */
 
 /**
@@ -91,7 +83,7 @@ import NftId from "../token/NftId.js";
 /**
  * Transfers a new Hedera™ crypto-currency token.
  */
-export default class TransferTransaction extends Transaction {
+export default class TransferTransaction extends AbstractTokenTransfer {
     /**
      * @param {object} [props]
      * @param {(TransferTokensInput)[]} [props.tokenTransfers]
@@ -103,49 +95,16 @@ export default class TransferTransaction extends Transaction {
 
         /**
          * @private
-         * @type {TokenTransfer[]}
-         */
-        this._tokenTransfers = [];
-
-        /**
-         * @private
          * @type {Transfer[]}
          */
         this._hbarTransfers = [];
 
-        /**
-         * @private
-         * @type {TokenNftTransfer[]}
-         */
-        this._nftTransfers = [];
-
         this._defaultMaxTransactionFee = new Hbar(1);
-
-        for (const transfer of props.tokenTransfers != null
-            ? props.tokenTransfers
-            : []) {
-            this.addTokenTransfer(
-                transfer.tokenId,
-                transfer.accountId,
-                transfer.amount,
-            );
-        }
 
         for (const transfer of props.hbarTransfers != null
             ? props.hbarTransfers
             : []) {
             this.addHbarTransfer(transfer.accountId, transfer.amount);
-        }
-
-        for (const transfer of props.nftTransfers != null
-            ? props.nftTransfers
-            : []) {
-            this.addNftTransfer(
-                transfer.tokenId,
-                transfer.serial,
-                transfer.sender,
-                transfer.recipient,
-            );
         }
     }
 
@@ -201,158 +160,6 @@ export default class TransferTransaction extends Transaction {
             nodeIds,
             bodies,
         );
-    }
-
-    /**
-     * @returns {TokenTransferMap}
-     */
-    get tokenTransfers() {
-        const map = new TokenTransferMap();
-
-        for (const transfer of this._tokenTransfers) {
-            let transferMap = map.get(transfer.tokenId);
-
-            if (transferMap != null) {
-                transferMap._set(transfer.accountId, transfer.amount);
-            } else {
-                transferMap = new TokenTransferAccountMap();
-                transferMap._set(transfer.accountId, transfer.amount);
-                map._set(transfer.tokenId, transferMap);
-            }
-        }
-
-        return map;
-    }
-
-    /**
-     * @param {TokenId | string} tokenId
-     * @param {AccountId | string} accountId
-     * @param {number | Long} amount
-     * @param {boolean} isApproved
-     * @returns {this}
-     */
-    _addTokenTransfer(tokenId, accountId, amount, isApproved) {
-        this._requireNotFrozen();
-
-        const token =
-            tokenId instanceof TokenId ? tokenId : TokenId.fromString(tokenId);
-        const account =
-            accountId instanceof AccountId
-                ? accountId
-                : AccountId.fromString(accountId);
-        const value = amount instanceof Long ? amount : Long.fromNumber(amount);
-
-        for (const tokenTransfer of this._tokenTransfers) {
-            if (
-                tokenTransfer.tokenId.compare(token) === 0 &&
-                tokenTransfer.accountId.compare(account) === 0
-            ) {
-                tokenTransfer.amount = tokenTransfer.amount.add(value);
-                tokenTransfer.expectedDecimals = null;
-                return this;
-            }
-        }
-
-        this._tokenTransfers.push(
-            new TokenTransfer({
-                tokenId,
-                accountId,
-                expectedDecimals: null,
-                amount,
-                isApproved,
-            }),
-        );
-
-        return this;
-    }
-
-    /**
-     * @param {TokenId | string} tokenId
-     * @param {AccountId | string} accountId
-     * @param {number | Long} amount
-     * @returns {this}
-     */
-    addTokenTransfer(tokenId, accountId, amount) {
-        return this._addTokenTransfer(tokenId, accountId, amount, false);
-    }
-
-    /**
-     * @param {TokenId | string} tokenId
-     * @param {AccountId | string} accountId
-     * @param {number | Long} amount
-     * @returns {this}
-     */
-    addApprovedTokenTransfer(tokenId, accountId, amount) {
-        return this._addTokenTransfer(tokenId, accountId, amount, true);
-    }
-
-    /**
-     * @param {TokenId | string} tokenId
-     * @param {AccountId | string} accountId
-     * @param {number | Long} amount
-     * @param {number} decimals
-     * @returns {this}
-     */
-    addTokenTransferWithDecimals(tokenId, accountId, amount, decimals) {
-        this._requireNotFrozen();
-
-        const token =
-            tokenId instanceof TokenId ? tokenId : TokenId.fromString(tokenId);
-        const account =
-            accountId instanceof AccountId
-                ? accountId
-                : AccountId.fromString(accountId);
-        const value = amount instanceof Long ? amount : Long.fromNumber(amount);
-
-        let found = false;
-
-        for (const tokenTransfer of this._tokenTransfers) {
-            if (tokenTransfer.tokenId.compare(token) === 0) {
-                if (
-                    tokenTransfer.expectedDecimals != null &&
-                    tokenTransfer.expectedDecimals !== decimals
-                ) {
-                    throw new Error("expected decimals mis-match");
-                } else {
-                    tokenTransfer.expectedDecimals = decimals;
-                }
-
-                if (tokenTransfer.accountId.compare(account) === 0) {
-                    tokenTransfer.amount = tokenTransfer.amount.add(value);
-                    tokenTransfer.expectedDecimals = decimals;
-                    found = true;
-                }
-            }
-        }
-
-        if (found) {
-            return this;
-        }
-
-        this._tokenTransfers.push(
-            new TokenTransfer({
-                tokenId,
-                accountId,
-                expectedDecimals: decimals,
-                amount,
-                isApproved: false,
-            }),
-        );
-
-        return this;
-    }
-
-    /**
-     * @returns {NullableTokenDecimalMap}
-     */
-    get tokenIdDecimals() {
-        const map = new NullableTokenDecimalMap();
-
-        for (const transfer of this._tokenTransfers) {
-            map._set(transfer.tokenId, transfer.expectedDecimals);
-        }
-
-        return map;
     }
 
     /**
@@ -453,186 +260,6 @@ export default class TransferTransaction extends Transaction {
     }
 
     /**
-     * @returns {TokenNftTransferMap}
-     */
-    get nftTransfers() {
-        const map = new TokenNftTransferMap();
-
-        for (const transfer of this._nftTransfers) {
-            const transferList = map.get(transfer.tokenId);
-
-            const nftTransfer = {
-                sender: transfer.senderAccountId,
-                recipient: transfer.receiverAccountId,
-                serial: transfer.serialNumber,
-                isApproved: transfer.isApproved,
-            };
-
-            if (transferList != null) {
-                transferList.push(nftTransfer);
-            } else {
-                map._set(transfer.tokenId, [nftTransfer]);
-            }
-        }
-
-        return map;
-    }
-
-    /**
-     * @param {boolean} isApproved
-     * @param {NftId | TokenId | string} tokenIdOrNftId
-     * @param {AccountId | string | Long | number} senderAccountIdOrSerialNumber
-     * @param {AccountId | string} receiverAccountIdOrSenderAccountId
-     * @param {(AccountId | string)=} receiver
-     * @returns {TransferTransaction}
-     */
-    _addNftTransfer(
-        isApproved,
-        tokenIdOrNftId,
-        senderAccountIdOrSerialNumber,
-        receiverAccountIdOrSenderAccountId,
-        receiver,
-    ) {
-        this._requireNotFrozen();
-
-        let nftId;
-        let senderAccountId;
-        let receiverAccountId;
-
-        if (tokenIdOrNftId instanceof NftId) {
-            nftId = tokenIdOrNftId;
-            senderAccountId =
-                typeof senderAccountIdOrSerialNumber === "string"
-                    ? AccountId.fromString(senderAccountIdOrSerialNumber)
-                    : /** @type {AccountId} */ (senderAccountIdOrSerialNumber);
-            receiverAccountId =
-                typeof receiverAccountIdOrSenderAccountId === "string"
-                    ? AccountId.fromString(receiverAccountIdOrSenderAccountId)
-                    : /** @type {AccountId} */ (
-                          receiverAccountIdOrSenderAccountId
-                      );
-        } else if (tokenIdOrNftId instanceof TokenId) {
-            nftId = new NftId(
-                tokenIdOrNftId,
-                /** @type {Long} */ (senderAccountIdOrSerialNumber),
-            );
-            senderAccountId =
-                typeof receiverAccountIdOrSenderAccountId === "string"
-                    ? AccountId.fromString(receiverAccountIdOrSenderAccountId)
-                    : /** @type {AccountId} */ (
-                          receiverAccountIdOrSenderAccountId
-                      );
-            receiverAccountId =
-                typeof receiver === "string"
-                    ? AccountId.fromString(receiver)
-                    : /** @type {AccountId} */ (receiver);
-        } else {
-            try {
-                nftId = NftId.fromString(tokenIdOrNftId);
-                senderAccountId =
-                    typeof senderAccountIdOrSerialNumber === "string"
-                        ? AccountId.fromString(senderAccountIdOrSerialNumber)
-                        : /** @type {AccountId} */ (
-                              senderAccountIdOrSerialNumber
-                          );
-                receiverAccountId =
-                    typeof receiverAccountIdOrSenderAccountId === "string"
-                        ? AccountId.fromString(
-                              receiverAccountIdOrSenderAccountId,
-                          )
-                        : /** @type {AccountId} */ (
-                              receiverAccountIdOrSenderAccountId
-                          );
-            } catch (_) {
-                const tokenId = TokenId.fromString(tokenIdOrNftId);
-                nftId = new NftId(
-                    tokenId,
-                    /** @type {Long} */ (senderAccountIdOrSerialNumber),
-                );
-                senderAccountId =
-                    typeof receiverAccountIdOrSenderAccountId === "string"
-                        ? AccountId.fromString(
-                              receiverAccountIdOrSenderAccountId,
-                          )
-                        : /** @type {AccountId} */ (
-                              receiverAccountIdOrSenderAccountId
-                          );
-                receiverAccountId =
-                    typeof receiver === "string"
-                        ? AccountId.fromString(receiver)
-                        : /** @type {AccountId} */ (receiver);
-            }
-        }
-
-        for (const nftTransfer of this._nftTransfers) {
-            if (
-                nftTransfer.tokenId.compare(nftId.tokenId) === 0 &&
-                nftTransfer.serialNumber.compare(nftId.serial) === 0
-            ) {
-                nftTransfer.senderAccountId = senderAccountId;
-                nftTransfer.receiverAccountId = receiverAccountId;
-                return this;
-            }
-        }
-
-        this._nftTransfers.push(
-            new TokenNftTransfer({
-                tokenId: nftId.tokenId,
-                serialNumber: nftId.serial,
-                senderAccountId,
-                receiverAccountId,
-                isApproved,
-            }),
-        );
-
-        return this;
-    }
-
-    /**
-     * @param {NftId | TokenId | string} tokenIdOrNftId
-     * @param {AccountId | string | Long | number} senderAccountIdOrSerialNumber
-     * @param {AccountId | string} receiverAccountIdOrSenderAccountId
-     * @param {(AccountId | string)=} receiver
-     * @returns {TransferTransaction}
-     */
-    addNftTransfer(
-        tokenIdOrNftId,
-        senderAccountIdOrSerialNumber,
-        receiverAccountIdOrSenderAccountId,
-        receiver,
-    ) {
-        return this._addNftTransfer(
-            false,
-            tokenIdOrNftId,
-            senderAccountIdOrSerialNumber,
-            receiverAccountIdOrSenderAccountId,
-            receiver,
-        );
-    }
-
-    /**
-     * @param {NftId | TokenId | string} tokenIdOrNftId
-     * @param {AccountId | string | Long | number} senderAccountIdOrSerialNumber
-     * @param {AccountId | string} receiverAccountIdOrSenderAccountId
-     * @param {(AccountId | string)=} receiver
-     * @returns {TransferTransaction}
-     */
-    addApprovedNftTransfer(
-        tokenIdOrNftId,
-        senderAccountIdOrSerialNumber,
-        receiverAccountIdOrSenderAccountId,
-        receiver,
-    ) {
-        return this._addNftTransfer(
-            true,
-            tokenIdOrNftId,
-            senderAccountIdOrSerialNumber,
-            receiverAccountIdOrSenderAccountId,
-            receiver,
-        );
-    }
-
-    /**
      * @deprecated - Use `addApprovedHbarTransfer()` instead
      * @param {AccountId | string} accountId
      * @param {boolean} isApproved
@@ -727,154 +354,7 @@ export default class TransferTransaction extends Transaction {
      * @returns {HashgraphProto.proto.ICryptoTransferTransactionBody}
      */
     _makeTransactionData() {
-        /** @type {{tokenId: TokenId; expectedDecimals: number | null; transfers: TokenTransfer[]; nftTransfers: TokenNftTransfer[];}[]} */
-        const tokenTransferList = [];
-
-        this._tokenTransfers.sort((a, b) => {
-            const compare = a.tokenId.compare(b.tokenId);
-
-            if (compare !== 0) {
-                return compare;
-            }
-
-            return a.accountId.compare(b.accountId);
-        });
-
-        this._nftTransfers.sort((a, b) => {
-            const senderComparision = a.senderAccountId.compare(
-                b.senderAccountId,
-            );
-            if (senderComparision != 0) {
-                return senderComparision;
-            }
-
-            const recipientComparision = a.receiverAccountId.compare(
-                b.receiverAccountId,
-            );
-            if (recipientComparision != 0) {
-                return recipientComparision;
-            }
-
-            return a.serialNumber.compare(b.serialNumber);
-        });
-
-        let i = 0;
-        let j = 0;
-        while (
-            i < this._tokenTransfers.length ||
-            j < this._nftTransfers.length
-        ) {
-            if (
-                i < this._tokenTransfers.length &&
-                j < this._nftTransfers.length
-            ) {
-                const iTokenId = this._tokenTransfers[i].tokenId;
-                const jTokenId = this._nftTransfers[j].tokenId;
-
-                const last =
-                    tokenTransferList.length > 0
-                        ? tokenTransferList[tokenTransferList.length - 1]
-                        : null;
-                const lastTokenId = last != null ? last.tokenId : null;
-
-                if (
-                    last != null &&
-                    lastTokenId != null &&
-                    lastTokenId.compare(iTokenId) === 0
-                ) {
-                    last.transfers.push(this._tokenTransfers[i++]);
-                    continue;
-                }
-
-                if (
-                    last != null &&
-                    lastTokenId != null &&
-                    lastTokenId.compare(jTokenId) === 0
-                ) {
-                    last.nftTransfers.push(this._nftTransfers[j++]);
-                    continue;
-                }
-
-                const result = iTokenId.compare(jTokenId);
-
-                if (result === 0) {
-                    tokenTransferList.push({
-                        tokenId: iTokenId,
-                        expectedDecimals:
-                            this._tokenTransfers[i].expectedDecimals,
-                        transfers: [this._tokenTransfers[i++]],
-                        nftTransfers: [this._nftTransfers[j++]],
-                    });
-                } else if (result < 0) {
-                    tokenTransferList.push({
-                        tokenId: iTokenId,
-                        expectedDecimals:
-                            this._tokenTransfers[i].expectedDecimals,
-                        transfers: [this._tokenTransfers[i++]],
-                        nftTransfers: [],
-                    });
-                } else {
-                    tokenTransferList.push({
-                        tokenId: jTokenId,
-                        expectedDecimals: null,
-                        transfers: [],
-                        nftTransfers: [this._nftTransfers[j++]],
-                    });
-                }
-            } else if (i < this._tokenTransfers.length) {
-                const iTokenId = this._tokenTransfers[i].tokenId;
-
-                let last;
-                for (const transfer of tokenTransferList) {
-                    if (transfer.tokenId.compare(iTokenId) === 0) {
-                        last = transfer;
-                    }
-                }
-                const lastTokenId = last != null ? last.tokenId : null;
-
-                if (
-                    last != null &&
-                    lastTokenId != null &&
-                    lastTokenId.compare(iTokenId) === 0
-                ) {
-                    last.transfers.push(this._tokenTransfers[i++]);
-                    continue;
-                }
-
-                tokenTransferList.push({
-                    tokenId: iTokenId,
-                    expectedDecimals: this._tokenTransfers[i].expectedDecimals,
-                    transfers: [this._tokenTransfers[i++]],
-                    nftTransfers: [],
-                });
-            } else if (j < this._nftTransfers.length) {
-                const jTokenId = this._nftTransfers[j].tokenId;
-
-                let last;
-                for (const transfer of tokenTransferList) {
-                    if (transfer.tokenId.compare(jTokenId) === 0) {
-                        last = transfer;
-                    }
-                }
-                const lastTokenId = last != null ? last.tokenId : null;
-
-                if (
-                    last != null &&
-                    lastTokenId != null &&
-                    lastTokenId.compare(jTokenId) === 0
-                ) {
-                    last.nftTransfers.push(this._nftTransfers[j++]);
-                    continue;
-                }
-
-                tokenTransferList.push({
-                    tokenId: jTokenId,
-                    expectedDecimals: null,
-                    transfers: [],
-                    nftTransfers: [this._nftTransfers[j++]],
-                });
-            }
-        }
+        const { tokenTransfers } = super._makeTransactionData();
 
         this._hbarTransfers.sort((a, b) => a.accountId.compare(b.accountId));
 
@@ -888,21 +368,7 @@ export default class TransferTransaction extends Transaction {
                     };
                 }),
             },
-            tokenTransfers: tokenTransferList.map((tokenTransfer) => {
-                return {
-                    token: tokenTransfer.tokenId._toProtobuf(),
-                    expectedDecimals:
-                        tokenTransfer.expectedDecimals != null
-                            ? { value: tokenTransfer.expectedDecimals }
-                            : null,
-                    transfers: tokenTransfer.transfers.map((transfer) =>
-                        transfer._toProtobuf(),
-                    ),
-                    nftTransfers: tokenTransfer.nftTransfers.map((transfer) =>
-                        transfer._toProtobuf(),
-                    ),
-                };
-            }),
+            tokenTransfers,
         };
     }
